apps/pyproject.toml
[project]
name = "genai"
version = "0.1.1"
description = "GenAI Learning"
authors = [
    {name = "Dilip Sharma",email = "diliplakshya@gmail.com"}
]
readme = "README.md"
requires-python = "3.12.12"
dependencies = [
    "streamlit (>=1.51.0,<2.0.0)",
    "openai-whisper (>=20250625,<20250626)",
    "fpdf (>=1.7.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false